requirements.txt
asgiref==3.10.0
crispy-bootstrap5==0.7
Django==5.2.7
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
et_xmlfile==2.0.0
gunicorn==21.2.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
requests>=2.31
